+++ lib/src/status_info.dart
@@ -1,5 +1,5 @@
-import 'package:zsdk/src/cause.dart';
-import 'package:zsdk/src/status.dart';
+import 'package:zsdk/src/enumerators/cause.dart';
+import 'package:zsdk/src/enumerators/status.dart';
 
 /// Created by luis901101 on 2020-01-07.
 class StatusInfo
@@ -8,6 +8,11 @@ class StatusInfo
     final Cause cause;
 
     StatusInfo(this.status, this.cause);
+
+    Map<String, dynamic> toMap() => <String, dynamic>{
+        'status': StatusUtils.get().nameOf(status),
+        'cause': CauseUtils.get().nameOf(cause),
+    };
 
     factory StatusInfo.fromMap(Map<dynamic, dynamic> map) =>
         map != null ?